+++ src/services/oidc.security.validation.ts
@@ -305,6 +305,10 @@ export class OidcSecurityValidation {
             return true;
         }
 
+        if (response_type === 'code') {
+            return true;
+        }
+
         this.loggerService.logWarning('module configure incorrect, invalid response_type:' + response_type);
         return false;
     }
@@ -329,10 +333,19 @@ export class OidcSecurityValidation {
     // access_token C2: Take the left- most half of the hash and base64url- encode it.
     // access_token C3: The value of at_hash in the ID Token MUST match the value produced in the previous step if at_hash
     // is present in the ID Token.
-    validate_id_token_at_hash(access_token: any, at_hash: any): boolean {
-        this.loggerService.logDebug('From the server:' + at_hash);
+    validate_id_token_at_hash(access_token: any, at_hash: any, isCodeFlow: boolean): boolean {
+        this.loggerService.logDebug('at_hash from the server:' + at_hash);
+
+        // The at_hash is optional for the code flow
+        if (isCodeFlow) {
+            if (!(at_hash as string)) {
+                this.loggerService.logDebug('Code Flow active, and no at_hash in the id_token, skipping check!');
+                return true;
+            }
+        }
+
         const testdata = this.generate_at_hash('' + access_token);
-        this.loggerService.logDebug('client validation not decoded:' + testdata);
+        this.loggerService.logDebug('at_hash client validation not decoded:' + testdata);
         if (testdata === (at_hash as string)) {
             return true; // isValid;
         } else {
@@ -350,6 +363,13 @@ export class OidcSecurityValidation {
         const hash = KJUR.crypto.Util.hashString(access_token, 'sha256');
         const first128bits = hash.substr(0, hash.length / 2);
         const testdata = hextob64u(first128bits);
+
+        return testdata;
+    }
+
+    generate_code_verifier(code_challenge: any): string {
+        const hash = KJUR.crypto.Util.hashString(code_challenge, 'sha256');
+        const testdata = hextob64u(hash);
 
         return testdata;
     }